re-organized utils/utils*.py
move sub-functions in pysarAp.py to utils/utils.py to prepare for --step option in pysarApp.py

split utils/utils.py into the following files for smaller files and more flexible dependencies:
utils/utils0.py
utils/utils1.py
objects/coord.py
legacy/utils.py

move insar_vs_gps.py from utils to objects

move standardize_trop_model() from utils/utils.py back to tropcor_pyaps.py

move timeseries2velocity.design_matrix() to objects.timeseries.get_design_matrix4average_velocity() for more clera dependency relationship because it's used in several place

update dependency graph description in __init__.py
