+++ pysar/objects/__init__.py
@@ -1,5 +1,15 @@
-#!/usr/bin/env python3
-
-from .pysarobj import *
 from .giantobj import *
+from .pysarobj import *
 from .ramp import *
+
+
+## Sub-module dependency graph:
+## Check pysar/__init__.py for more comprehensiave graph
+# Level 0 modules:
+# giantobj
+# pysarobj
+# ramp
+# sensor
+# Level 1 modules:
+# conncomp (objects/ramp)
+#
