+++ pysar/objects/pysarobj.py
@@ -385,6 +385,25 @@ class timeseries:
             for d, pbase in zip(date6_list, pbase_list):
                 f.write('{}\t{}\n'.format(d, pbase))
         return out_file
+
+    # Functions for Unwrap error correction
+    @staticmethod
+    def get_design_matrix4average_velocity(date_list):
+        """design matrix/function model of linear velocity estimation
+        Parameters: date_list : list of string in YYYYMMDD format
+        Returns:    A : 2D array of int in size of (numDate, 2)
+        """
+        # convert list of YYYYMMDD into array of diff year in float
+        dt_list = [dt.strptime(i, '%Y%m%d') for i in date_list]
+        yr_list = [i.year + (i.timetuple().tm_yday - 1) / 365.25 for i in dt_list]
+        yr_diff = np.array(yr_list)
+        yr_diff -= yr_diff[0]
+
+        #for precision, use float32 in 0.1 yr, or float64 in 2015.1 yr format
+        A = np.ones([len(date_list), 2], dtype=np.float32)
+        A[:, 0] = yr_diff
+        return A
+
 ################################ timeseries class end ##################################
